refactor: folder restructuring v3
