+++ src/components/SelectSocialProvider.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import {
   Select,
   SelectContent,
