+++ src/components/UserLookupForm.tsx
@@ -9,8 +9,6 @@ import { UserContext } from "@/context/user-provider";
 import { SupportedSocialNetworkIds, UserLookupBy } from "@/types";
 import { useModifyQueryParams } from "@/libs/hooks/useModifyQueryParams";
 import { LookupInput } from "./ui/lookup-input";
-import { useUserLookupBy } from "@/libs/hooks/useUserLookupBy";
-import { FormSubmissionLoader } from "./FormSubmissionLoader";
 
 /**
  * Form to lookup a user's wallet based on query params and client-side routing e.g. in "/home" page
@@ -64,80 +62,6 @@ export const UserLookupClientForm = () => {
         >
           Look up wallet
         </Button>
-      </form>
-    </div>
-  );
-};
-
-export const HeaderLookupForm = () => {
-  const [searchValue, setSearchValue] = useState("");
-  const [searchProvider, setSearchProvider] =
-    useState<SupportedSocialNetworkIds>("twitter");
-  const { chain } = useContext(UserContext);
-  const lookupProviderDetails = getSupportedLookupNetworks()[searchProvider];
-  const pathname = usePathname();
-
-  const { push } = useRouter();
-
-  const handleOnSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    push(`/${searchProvider}:${searchValue}/${chain}`);
-  };
-  return (
-    <form
-      className={cn("flex w-full flex-1 items-center gap-2", {
-        hidden: pathname === "/",
-      })}
-      onSubmit={handleOnSubmit}
-    >
-      <LookupInput
-        onInputChange={(e) => setSearchValue(e.target.value)}
-        onSelectChange={(value) =>
-          setSearchProvider(value as SupportedSocialNetworkIds)
-        }
-        placeholder={lookupProviderDetails.placeholder}
-        className="w-full min-w-[260px]"
-      />
-    </form>
-  );
-};
-
-/**
- * Form to lookup a user's wallet based on serverside form submission e.g. in "Send" modal
- */
-export const UserLookupServerForm = ({
-  by = "default",
-}: {
-  by: UserLookupBy;
-}) => {
-  const { content, formAction, formRef, state } = useUserLookupBy({ by });
-  const isDefault = by === "default";
-
-  return (
-    <div className="flex w-full">
-      <form
-        action={formAction}
-        ref={formRef}
-        className="relative flex flex-1 items-center"
-      >
-        <LookupInput
-          onInputChange={content.onInputChange}
-          onSelectChange={content.onInputChange}
-          className={cn({ "pl-24": by !== "default" })}
-          placeholder={content.placeholder}
-          leftButton={
-            !isDefault ? (
-              <div className="flex items-center rounded-lg border-[0.5px] border-gray-800 bg-gray-900 p-2.5 py-1 text-gray-600">
-                <div>{content.btnTitle}</div>
-                <input type="hidden" name="provider" value={by} />
-              </div>
-            ) : undefined
-          }
-        />
-        <FormSubmissionLoader className="absolute right-4 text-gray-300" />
-        <p className="absolute -bottom-6 left-2 text-sm text-red-600">
-          {state.errorMessage}
-        </p>
       </form>
     </div>
   );
