+++ src/components/layout/Nav.tsx
@@ -3,10 +3,10 @@ import { Suspense } from "react";
 import { NavSignIn } from "./NavSignIn";
 import Image from "next/image";
 import walletLogo from "../../../public/wallet_logo.svg";
-import { HeaderLookupForm } from "../UserLookupForm";
 import Link from "next/link";
 import { HeaderSkeleton } from "../Skeleton";
 import AlphaBanner from "./AlphaBanner";
+import { HeaderLookupForm } from "./HeaderLookupForm";
 
 export default function Nav() {
   return (
